refactor node availability to HostNode::Update mutation
