+++ server/spec/mutations/host_nodes/update_spec.rb
@@ -4,24 +4,88 @@ describe HostNodes::Update, celluloid: true do
   let(:node) { HostNode.create!(name: 'node-1', grid: grid, node_id: 'AA') }
 
   describe '#run' do
-    it 'updates node labels' do
-      node.labels = []
-      labels = ['foo=bar', 'bar=baz']
-      expect {
-        described_class.new(
+    context 'labels' do
+      it 'updates node labels' do
+        node.labels = []
+        labels = ['foo=bar', 'bar=baz']
+        expect {
+          described_class.new(
+            host_node: node,
+            labels: labels
+          ).run
+        }.to change{ node.labels }.from([]).to(labels)
+      end
+
+      it 'does not update lables if nil labels given' do
+        node.labels = ['foo=bar']
+        expect {
+          described_class.new(
+            host_node: node,
+            labels: nil
+          ).run
+        }.not_to change{ node.labels }
+      end
+
+      it 'notifies grid nodes' do
+        mutation = described_class.new(
           host_node: node,
-          labels: labels
-        ).run
-      }.to change{ node.labels }.from([]).to(labels)
+          labels: []
+        )
+        expect(mutation).to receive(:notify_grid).once.with(grid)
+        mutation.run
+      end
     end
 
-    it 'notifies grid nodes' do
-      mutation = described_class.new(
-        host_node: node,
-        labels: []
-      )
-      expect(mutation).to receive(:notify_grid).once.with(grid)
-      mutation.run
+    context 'availability' do
+      context 'drain' do
+        it 'updates availability to drain' do
+          expect {
+            outcome = described_class.new(
+              host_node: node,
+              availability: 'drain',
+              labels: []
+            ).run
+            expect(outcome.success?).to be_truthy
+          }.to change{ node.reload.availability }.from('active').to('drain')
+        end
+
+        it 'starts re-scheduling when draining' do
+          mutation = described_class.new(
+            host_node: node,
+            availability: 'drain',
+            labels: []
+          )
+          expect(mutation).to receive(:re_deploy_needed_services)
+          expect(mutation).to receive(:stop_stateful_services)
+
+          mutation.run
+        end
+      end
+
+      context 'active' do
+        it 'updates availability to active' do
+          node.availability = 'drain'
+          expect {
+            described_class.new(
+              host_node: node,
+              availability: 'active',
+              labels: []
+            ).run
+          }.to change{ node.availability }.from('drain').to('active')
+        end
+      end
+
+      it 'triggers actions only when availability changes' do
+        node.availability = 'drain'
+        expect(node).not_to receive(:set)
+        expect {
+          described_class.new(
+            host_node: node,
+            availability: 'drain',
+            labels: []
+          ).run
+        }.not_to change{ node }
+      end
     end
   end
 end
